requirements.txt
aiobotocore==2.7.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aioitertools==0.12.0
aiosignal==1.3.2
altair==5.5.0
anyio==4.8.0
async-timeout==5.0.1
atlassian-python-api==3.41.19
attrs==25.1.0
beautifulsoup4==4.13.3
botocore==1.31.64
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
datasets==2.16.1
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2023.10.0
gcsfs==2023.10.0
google-api-core==2.24.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.2
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
gradio==3.50.2
gradio_client==0.6.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lxml==5.3.1
MarkupSafe==2.1.5
matplotlib==3.9.4
mmh3==4.1.0
multidict==6.1.0
multiprocess==0.70.15
narwhals==1.29.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.65.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pinecone-client==3.2.2
pinecone-datasets==0.7.0
pinecone-text==0.9.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==11.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.10.21
pydub==0.25.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
rsa==4.9
s3fs==2023.10.0
scikit-learn==1.5.0
scipy==1.13.1
semantic-version==2.10.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.0
tqdm==4.67.1
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
uvicorn==0.34.0
websockets==11.0.0
wget==3.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
